requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
email-validator==2.1.0.post1

python-multipart==0.0.12
pandas==3.0.1
openpyxl==3.1.5
python-docx==1.1.2
pypdf==5.1.0
pdfplumber==0.11.4
python-dotenv==1.0.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary,pool]==3.2.5

# === SECURITY (M4A-F) ===
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
redis==5.2.1
filetype==1.2.0
werkzeug==3.1.3

pytest>=8.0.0
pytest-cov>=4.1.0
httpx==0.27.0

# === RESILIENCE (M4D) ===
tenacity==9.0.0
pybreaker==1.2.0
llama-cloud>=1.5.0
